Colocate messages and types
