+++ react/components/CategoryItem.tsx
@@ -15,7 +15,7 @@ class CategoryItem extends Component<CategoryItemProps, State> {
   }
 
   public render() {
-    const { categoryId, text, typography, isHovered } = this.props
+    const { categoryId, text, typography, isHovered, highlight } = this.props
 
     if (this.state.hasError) {
       return null
@@ -35,6 +35,7 @@ class CategoryItem extends Component<CategoryItemProps, State> {
 
           return (
             <StyledLink 
+              highlight={highlight}
               isHovered={isHovered}
               typography={typography}
               to={data.category.href}>
@@ -49,7 +50,13 @@ class CategoryItem extends Component<CategoryItemProps, State> {
 
 export interface CategoryItemProps extends CategoryItemSchema {
   isHovered: boolean
+  highlight: boolean
   typography?: string
+}
+
+export interface CategoryItemSchema {
+  categoryId: number
+  text: string
 }
 
 interface State {
