--- react/modules/messages.ts
@@ -1,116 +1,0 @@
-import { defineMessages } from 'react-intl'
-
-export default defineMessages({
-  categoryIdTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.item.params.categoryId.title',
-  },
-  categoryTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.item.category.title',
-  },
-  customTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.item.custom.title',
-  },
-  customTypeTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.item.params.type.title',
-  },
-  externalTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.item.params.external.title',
-  },
-  hasSubMenuTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.item.hasSubMenu.title',
-  },
-  highlightTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.item.highlight.title',
-  },
-  horizontalLabel: {
-    defaultMessage: '',
-    id: 'editor.menu.orientation.horizontal.label',
-  },
-  hrefTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.item.params.href.title',
-  },
-  iconIdTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.item.iconId.title',
-  },
-  internalTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.item.params.internal.title',
-  },
-  itemIdTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.item.id.title',
-  },
-  itemsDescription: {
-    defaultMessage: '',
-    id: 'editor.menu.items.description',
-  },
-  itemsTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.items.title',
-  },
-  menuDescription: {
-    defaultMessage: '',
-    id: 'editor.future-menu.description',
-  },
-  menuTitle: {
-    defaultMessage: '',
-    id: 'editor.future-menu.title',
-  },
-  noFollowTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.item.params.noFollow.title',
-  },
-  orientationDescription: {
-    defaultMessage: '',
-    id: 'editor.menu.typography.description',
-  },
-  orientationTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.orientation.title',
-  },
-  paramsTextTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.item.params.text.title',
-  },
-  paramsTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.item.params.title',
-  },
-  submenuTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.submenu.title',
-  },
-  submenuWidthTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.item.submenuWidth.title',
-  },
-  tagTitleTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.item.params.tagTitle.title',
-  },
-  typeTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.item.type.title',
-  },
-  typographyDescription: {
-    defaultMessage: '',
-    id: 'editor.menu.typography.description',
-  },
-  typographyTitle: {
-    defaultMessage: '',
-    id: 'editor.menu.typography.title',
-  },
-  verticalLabel: {
-    defaultMessage: '',
-    id: 'editor.menu.orientation.vertical.label',
-  },
-})
